Stupid program that allow reading the Gyro/Acc
Yes, it is very stupid. Plug the USB in your PC and have a look at the wonderful
prompt that will remind you of the good ol' 70s. Enter 0/1 (gyro/acc), then a
space and an address in hex and it will print the value read. Input is validated
and when an error occur, a very user-friendly message appears (that is, when
compared to computers of the 70s).